+++ start.sh
@@ -1,2 +1,5 @@
-PORT=8008 DATABASE_URL=jdbc:sqlite:db/dev.sqlite java -jar target/where-is-me-0.2.2-standalone.jar
+#!/bin/sh
+PORT=8008 \
+DATABASE_URL=jdbc:sqlite:db/dev.sqlite \
+  java -jar target/where-is-me-0.2.2-standalone.jar